+++ README.md
@@ -45,10 +45,10 @@
 
 ### 요구사항
 
-- [ ] 본인만의 기준으로 일관된 코드를 작성해주세요.
-- [ ] 기능 단위로 나누어 커밋을 해주세요.
-- [ ] 재사용성을 고려하여 컴포넌트, 비즈니스 로직을 구현해요.
-- [ ] 테마 상품 목록 페이지
+- [x] 본인만의 기준으로 일관된 코드를 작성해주세요.
+- [x] 기능 단위로 나누어 커밋을 해주세요.
+- [x] 재사용성을 고려하여 컴포넌트, 비즈니스 로직을 구현해요.
+- [x] 테마 상품 목록 페이지
   - [x] 선물하기 홈 > 선물 테마 섹션의 아이템을 클릭하면 테마 상품 목록 페이지로 연결돼요.
   - [x] **`/api/themes/:themeId/info`** API를 사용하여 선물 테마 섹션의 히어로 영역을 구현해요.
   - [x] 만약 테마 정보를 가져오는 API가 404 에러가 발생하면 선물하기 홈 페이지로 연결해요.
