+++ src/components/ThemeSection/ThemeProductList.tsx
@@ -1,92 +1,15 @@
-import { useEffect, useState, useRef, useCallback } from 'react';
 import { useParams } from 'react-router-dom';
-import axios from 'axios';
 import styled from '@emotion/styled';
-import { ERROR_MESSAGES } from '@/constants/validation';
 import ProductCard from '@/components/RankingSection/ProductCard';
 import CardGrid from '@/components/common/CardGrid';
-import { getThemeProductsUrl } from '@/constants/api';
+import { ERROR_MESSAGES } from '@/constants/validation';
 import { loading } from '@/components/common/Loading';
-
-interface Product {
-  id: number;
-  name: string;
-  price: {
-    basicPrice: number;
-    sellingPrice: number;
-    discountRate: number;
-  };
-  imageURL: string;
-  brandInfo: {
-    id: number;
-    name: string;
-    imageURL: string;
-  };
-}
-
-const LIMIT = 10;
+import { useThemeProducts } from '@/hooks/useThemeProducts';
 
 const ThemeProductList = () => {
   const { themeId } = useParams<{ themeId: string }>();
-  const [products, setProducts] = useState<Product[]>([]);
-  const [cursor, setCursor] = useState(0);
-  const [hasMore, setHasMore] = useState(true);
-  const [pending, setPending] = useState(false);
-  const [error, setError] = useState(false);
-
-  const observerRef = useRef<HTMLDivElement | null>(null);
-
-  const fetchProducts = useCallback(async () => {
-    if (!themeId || !hasMore || pending) return;
-    setPending(true);
-    setError(false);
-    try {
-      const res = await axios.get<{
-        data: { list: Product[]; hasMoreList: boolean; cursor: number };
-      }>(getThemeProductsUrl(themeId), {
-        params: { cursor, limit: LIMIT },
-      });
-
-      const {
-        list: newProducts,
-        hasMoreList,
-        cursor: nextCursor,
-      } = res.data.data;
-
-      setProducts(prev => {
-        const merged = [...prev, ...newProducts];
-        const unique = Array.from(new Map(merged.map(p => [p.id, p])).values());
-        return unique;
-      });
-      setHasMore(hasMoreList);
-      setCursor(nextCursor);
-    } catch {
-      setError(true);
-    } finally {
-      setPending(false);
-    }
-  }, [themeId, cursor, hasMore, pending]);
-
-  useEffect(() => {
-    fetchProducts();
-  }, [fetchProducts]);
-
-  useEffect(() => {
-    if (!hasMore || pending) return;
-    const observer = new IntersectionObserver(
-      entries => {
-        if (entries[0].isIntersecting) {
-          fetchProducts();
-        }
-      },
-      { threshold: 1 }
-    );
-    const el = observerRef.current;
-    if (el) observer.observe(el);
-    return () => {
-      if (el) observer.unobserve(el);
-    };
-  }, [fetchProducts, hasMore, pending]);
+  const { products, pending, error, hasMore, observerRef } =
+    useThemeProducts(themeId);
 
   if (error)
     return <ErrorText>{ERROR_MESSAGES.FAILED_TO_LOAD_PRODUCTS}</ErrorText>;
